requirements.txt
streamlit
pandas
matplotlib
openai
python-dotenv
google-generativeai